+++ pkgs/native_assets_builder/test_data/wrong_build_output_3/hook/build.dart
@@ -12,8 +12,7 @@ void main(List<String> args) async {
   final inputPath = getInputArgument(args);
   final buildInput = BuildInput(
       json.decode(File(inputPath).readAsStringSync()) as Map<String, Object?>);
-  await File.fromUri(buildInput.outputDirectory.resolve('build_output.json'))
-      .writeAsString(_rightContents);
+  await File.fromUri(buildInput.outputFile).writeAsString(_rightContents);
   exit(1);
 }
 
@@ -22,5 +21,5 @@ const _rightContents = '''{
   "encodedAssets": [],
   "dependencies": [],
   "metadata": {},
-  "version": "1.7.0"
+  "version": "1.8.0"
 }''';