+++ pkgs/native_assets_cli/lib/src/api/link.dart
@@ -51,8 +51,7 @@ Future<void> link(
   if (errors.isEmpty) {
     final jsonOutput =
         const JsonEncoder().fuse(const Utf8Encoder()).convert(output.json);
-    await File.fromUri(input.outputDirectory.resolve('link_output.json'))
-        .writeAsBytes(jsonOutput);
+    await File.fromUri(input.outputFile).writeAsBytes(jsonOutput);
   } else {
     final message = [
       'The output contained unsupported output:',