+++ pkgs/native_assets_cli/lib/src/model/hook.dart
@@ -16,5 +16,5 @@ enum Hook {
 
   String get scriptName => '$_scriptName.dart';
 
-  String get outputName => '${_scriptName}_output.json';
+  String get outputNameDeprecated => '${_scriptName}_output.json';
 }