+++ pkgs/native_assets_cli/test/example/local_asset_test.dart
@@ -31,6 +31,7 @@ void main() async {
   for (final dryRun in [true, false]) {
     final testSuffix = dryRun ? ' dry_run' : '';
     test('local_asset build$testSuffix', () async {
+      final buildOutputUri = tempUri.resolve('build_output.json');
       final testTempUri = tempUri.resolve('test1/');
       await Directory.fromUri(testTempUri).create();
       final outputDirectory = tempUri.resolve('out/');
@@ -45,6 +46,7 @@ void main() async {
         ..setupShared(
           packageRoot: testPackageUri,
           packageName: name,
+          outputFile: buildOutputUri,
           outputDirectory: outputDirectory,
           outputDirectoryShared: outputDirectoryShared,
         )
@@ -79,7 +81,6 @@ void main() async {
       }
       expect(processResult.exitCode, 0);
 
-      final buildOutputUri = outputDirectory.resolve('build_output.json');
       final buildOutput = BuildOutput(
           json.decode(await File.fromUri(buildOutputUri).readAsString())
               as Map<String, Object?>);