+++ pkgs/native_toolchain_c/test/cbuilder/compiler_resolver_test.dart
@@ -47,6 +47,7 @@ void main() {
         ..setupShared(
           packageName: 'dummy',
           packageRoot: tempUri,
+          outputFile: tempUri.resolve('../output.json'),
           outputDirectory: tempUri,
           outputDirectoryShared: tempUri2,
         )
@@ -92,6 +93,7 @@ void main() {
       ..setupShared(
         packageName: 'dummy',
         packageRoot: tempUri,
+        outputFile: tempUri.resolve('../output.json'),
         outputDirectoryShared: tempUri2,
         outputDirectory: tempUri,
       )